checked profile response before deleting user account

+++ src/auth/auth.service.ts
@@ -319,20 +319,28 @@ export class AuthService {
           ),
         );
       }
-
-      const profileID = user.profileID;
-      await this.userRepo.delete({ email });
       //  send Delete request to the profile service to delete the user profile
       // Axios
       const coreUrl = `${this.configService.get<string>(
         configConstant.baseUrls.coreService,
-      )}/profile/${profileID}`;
+      )}/profile/${user.profileID}`;
 
-      await this.httpService.axiosRef({
+      const profileResponse = await this.httpService.axiosRef({
         method: 'delete',
         url: coreUrl,
       });
 
+      if (profileResponse.data.status !== 200) {
+        throw new BadRequestException(
+          ZaLaResponse.BadRequest(
+            'Profile Not deleted',
+            'Error occured while deleting user Profile',
+            '400',
+          ),
+        );
+      }
+
+      return await this.userRepo.delete({ email });
     } catch (error) {
       throw new BadRequestException(
         ZaLaResponse.BadRequest(error.name, error.message, error.status),
